Fix formal/actual argument inconsistency in complex_math_output.c by improving type inference in parser
Agent-Logs-Url: https://github.com/snehadpatel/CompilerDesign/sessions/7099f839-6f81-4353-b7d5-f6cf722ea7ca

+++ outputs/complex_math_output.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 
-int complex_calc(int a, int b) {
-    int res = (a + b) * (a - b) / 2;
+float complex_calc(float a, float b) {
+    float res = (a + b) * (a - b) / 2;
     return res;
 }
 int main() {
     float val_a = 15.5;
     float val_b = 4.5;
-    int ans = complex_calc(val_a, val_b);
+    float ans = complex_calc(val_a, val_b);
     printf("Calculation Result:\n");
-    printf("%d\n", ans);
+    printf("%g\n", ans);
     return 0;
 }
